requirements.txt
aiohttp
asyncio
numpy
pandas
rich
requests
python-dotenv
